// lib/api/generated/model/dailyQuizResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { QuizSessionResponseDto } from './quizSessionResponseDto';
import type { QuizDto } from './quizDto';

export interface DailyQuizResponseDto {
  session: QuizSessionResponseDto;
  quizzes: QuizDto[];
  /** 오늘의 퀴즈 완료 여부 */
  isCompleted: boolean;
}