// lib/api/generated/model/startQuizSessionDtoType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */

export type StartQuizSessionDtoType = typeof StartQuizSessionDtoType[keyof typeof StartQuizSessionDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StartQuizSessionDtoType = {
  EPISODE: 'EPISODE',
  PRACTICE: 'PRACTICE',
  DAILY_QUIZ: 'DAILY_QUIZ',
} as const;